👌 IMPROVE: redux code refactor & restructuring
