--- src/app/actions/productActions.ts
@@ -1,12 +1,0 @@
-import { createAction } from '@reduxjs/toolkit';
-import { Product } from '../../types';
-
-export const fetchProductsRequested = createAction(
-  'products/fetchProductsRequested',
-);
-export const fetchProductsSucceeded = createAction<Product[]>(
-  'products/fetchProductsSucceeded',
-);
-export const fetchProductsFailed = createAction<string>(
-  'products/fetchProductsFailed',
-);
