+++ src/app/store/reducer.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from "@reduxjs/toolkit";
+import productReducer from "./slices/productSlice";
+
+const rootReducer = combineReducers({
+  products: productReducer,
+});
+
+export default rootReducer;
