+++ src/app/store/sagas/productSaga.ts
@@ -1,23 +1,25 @@
-import { takeEvery, put, call } from 'redux-saga/effects';
-import { axiosInstance } from '../../utils/http';
+import { takeEvery, put, call } from "redux-saga/effects";
+import { axiosInstance } from "../../../utils/http";
 import {
   fetchProductsRequested,
   fetchProductsSucceeded,
   fetchProductsFailed,
-} from '../actions/productActions';
-import { Product } from '../../types';
-import { AxiosResponse } from 'axios';
+} from "../slices/productSlice";
+import { Product } from "../../../types";
+import { AxiosResponse } from "axios";
 
 function* fetchProducts(): Generator {
   try {
-    const response = yield call(() => axiosInstance.get(''));
+    const response = yield call(() => axiosInstance.get(""));
     const products: Product[] = (response as AxiosResponse<Product[]>).data;
     yield put(fetchProductsSucceeded(products));
   } catch (error) {
+    console.log("Error fetching products", error);
+
     if (error instanceof Error) {
       yield put(fetchProductsFailed(error.message));
     } else {
-      yield put(fetchProductsFailed('An unknown error occurred'));
+      yield put(fetchProductsFailed("An unknown error occurred"));
     }
   }
 }
